+++ src/Rhetos.CommonConcepts/Dom.DefaultConcepts/GenericFilterHelper.cs
@@ -50,6 +50,33 @@ namespace Rhetos.Dom.DefaultConcepts
             _logger = logProvider.GetLogger(GetType().Name);
         }
 
+        /// <summary>Generic property filter operation.</summary>
+        enum Op
+        {
+            Equals /*legacy Equal*/, NotEquals /*legacy NotEqual*/,
+            Greater, GreaterEqual, Less, LessEqual,
+            StartsWith, EndsWith, Contains, NotContains,
+            DateIn, DateNotIn,
+            In, NotIn, ContainsAny, StartsWithAny, EndsWithAny
+        };
+
+        static Op GetFilterOperation(PropertyFilter filter)
+        {
+            if (filter.Operation is null or "")
+                throw new ClientException($"Generic filter operation is not specified on property '{filter.Property}'.");
+
+            // Operations 'equal' and 'notequal' are supported for backward compatibility.
+            if (string.Equals(filter.Operation, "Equal", StringComparison.OrdinalIgnoreCase))
+                filter.Operation = Op.Equals.ToString();
+            if (string.Equals(filter.Operation, "NotEqual", StringComparison.OrdinalIgnoreCase))
+                filter.Operation = Op.NotEquals.ToString();
+
+            var filterOperations = Enum.GetValues<Op>().Where(fo => fo.ToString().Equals(filter.Operation, StringComparison.OrdinalIgnoreCase)).ToList();
+            if (filterOperations.Count == 0)
+                throw new ClientException($"Unsupported generic filter operation '{filter.Operation}'.");
+            return filterOperations.Single();
+        }
+
         //================================================================
         #region Property filters
 
@@ -87,19 +114,18 @@ namespace Rhetos.Dom.DefaultConcepts
 
                 bool propertyIsNullableValueType = memberAccess.Type.IsGenericType && memberAccess.Type.GetGenericTypeDefinition() == typeof(Nullable<>);
                 Type propertyBasicType = propertyIsNullableValueType ? memberAccess.Type.GetGenericArguments().Single() : memberAccess.Type;
+                Op filterOperation = GetFilterOperation(filter);
 
                 ConstantExpression constant;
 
-                switch (filter.Operation.ToLowerInvariant())
+                switch (filterOperation)
                 {
-                    case "equals":
-                    case "equal": // For backward compatibility.
-                    case "notequals":
-                    case "notequal": // For backward compatibility.
-                    case "greater":
-                    case "greaterequal":
-                    case "less":
-                    case "lessequal":
+                    case Op.Equals:
+                    case Op.NotEquals:
+                    case Op.Greater:
+                    case Op.GreaterEqual:
+                    case Op.Less:
+                    case Op.LessEqual:
                         {
                             // Constant value should be of same type as the member it is compared to.
                             object convertedValue;
@@ -129,21 +155,21 @@ namespace Rhetos.Dom.DefaultConcepts
                             break;
                         }
 
-                    case "startswith":
-                    case "endswith":
-                    case "contains":
-                    case "notcontains":
+                    case Op.StartsWith:
+                    case Op.EndsWith:
+                    case Op.Contains:
+                    case Op.NotContains:
                         // Constant value should be string.
                         constant = Expression.Constant(filter.Value.ToString(), typeof(string));
                         break;
 
-                    case "datein":
-                    case "datenotin":
+                    case Op.DateIn:
+                    case Op.DateNotIn:
                         constant = null;
                         break;
 
-                    case "in":
-                    case "notin":
+                    case Op.In:
+                    case Op.NotIn:
                         {
                             if (filter.Value == null)
                                 throw new ClientException($"Invalid generic filter parameter for operation '{filter.Operation}' on {propertyBasicType.Name} property '{filter.Property}'."
@@ -205,10 +231,9 @@ namespace Rhetos.Dom.DefaultConcepts
                 }
 
                 Expression expression;
-                switch (filter.Operation.ToLowerInvariant())
+                switch (filterOperation)
                 {
-                    case "equals":
-                    case "equal":
+                    case Op.Equals:
                         if (propertyBasicType == typeof(Guid) && constant.Value is Guid constantIdEquals)
                         {
                             // Using a different expression instead of the constant, to force Entity Framework to
@@ -225,8 +250,7 @@ namespace Rhetos.Dom.DefaultConcepts
                             expression = Expression.Equal(memberAccess, constant);
                         break;
 
-                    case "notequals":
-                    case "notequal":
+                    case Op.NotEquals:
                         if (propertyBasicType == typeof(Guid) && constant.Value is Guid constantIdNotEquals)
                         {
                             // Using a different expression instead of the constant, to force Entity Framework to
@@ -243,7 +267,7 @@ namespace Rhetos.Dom.DefaultConcepts
                             expression = Expression.NotEqual(memberAccess, constant);
                         break;
 
-                    case "greater":
+                    case Op.Greater:
                         if (propertyBasicType == typeof(string))
                             expression = Expression.Call(_ormUtility.IsGreaterThanMethod, memberAccess, constant);
                         else if (propertyBasicType == typeof(Guid))
@@ -251,7 +275,7 @@ namespace Rhetos.Dom.DefaultConcepts
                         else expression = Expression.GreaterThan(memberAccess, constant);
                         break;
 
-                    case "greaterequal":
+                    case Op.GreaterEqual:
                         if (propertyBasicType == typeof(string))
                             expression = Expression.Call(_ormUtility.IsGreaterThanOrEqualMethod, memberAccess, constant);
                         else if (propertyBasicType == typeof(Guid))
@@ -259,7 +283,7 @@ namespace Rhetos.Dom.DefaultConcepts
                         else expression = Expression.GreaterThanOrEqual(memberAccess, constant);
                         break;
 
-                    case "less":
+                    case Op.Less:
                         if (propertyBasicType == typeof(string))
                             expression = Expression.Call(_ormUtility.IsLessThanMethod, memberAccess, constant);
                         else if (propertyBasicType == typeof(Guid))
@@ -267,7 +291,7 @@ namespace Rhetos.Dom.DefaultConcepts
                         else expression = Expression.LessThan(memberAccess, constant);
                         break;
 
-                    case "lessequal":
+                    case Op.LessEqual:
                         if (propertyBasicType == typeof(string))
                             expression = Expression.Call(_ormUtility.IsLessThanOrEqualMethod, memberAccess, constant);
                         else if (propertyBasicType == typeof(Guid))
@@ -275,44 +299,52 @@ namespace Rhetos.Dom.DefaultConcepts
                         else expression = Expression.LessThanOrEqual(memberAccess, constant);
                         break;
 
-                    case "startswith":
-                    case "endswith":
+                    case Op.StartsWith:
+                    case Op.EndsWith:
+                    case Op.Contains:
+                    case Op.NotContains:
                         {
-                            Expression stringMember;
-                            if (propertyBasicType == typeof(string))
-                                stringMember = memberAccess;
-                            else
+                            MethodInfo dbMethod = filterOperation switch
                             {
-                                if (propertyBasicType != typeof(int))
-                                    throw new FrameworkException("Generic filter operation '" + filter.Operation + "' is not supported on property type '" + propertyBasicType.Name + "'.");
-                                stringMember = Expression.Call(_ormUtility.CastToStringMethod, memberAccess);
-                            }
-                            var dbMethod = filter.Operation.Equals("startswith", StringComparison.OrdinalIgnoreCase)
-                                ? _ormUtility.StartsWithCaseInsensitiveMethod
-                                : _ormUtility.EndsWithCaseInsensitiveMethod;
-                            expression = Expression.Call(dbMethod, stringMember, constant);
-                            break;
-                        }
-                    case "contains":
-                    case "notcontains":
-                        {
-                            Expression stringMember;
-                            if (propertyBasicType == typeof(string))
-                                stringMember = memberAccess;
-                            else
-                            {
-                                if (propertyBasicType != typeof(int))
-                                    throw new FrameworkException("Generic filter operation '" + filter.Operation + "' is not supported on property type '" + propertyBasicType.Name + "'.");
-                                stringMember = Expression.Call(_ormUtility.CastToStringMethod, memberAccess);
-                            }
-                            expression = Expression.Call(_ormUtility.ContainsCaseInsensitiveMethod, stringMember, constant);
+                                Op.StartsWith => _ormUtility.StartsWithCaseInsensitiveMethod,
+                                Op.EndsWith => _ormUtility.EndsWithCaseInsensitiveMethod,
+                                _ => _ormUtility.ContainsCaseInsensitiveMethod
+                            };
+                            Expression stringMemberAccess = GetStringMemberAccess(memberAccess, propertyBasicType, filter.Operation);
+                            expression = Expression.Call(dbMethod, stringMemberAccess, constant);
 
-                            if (filter.Operation.Equals("notcontains", StringComparison.OrdinalIgnoreCase))
+                            if (filterOperation == Op.NotContains)
                                 expression = Expression.Not(expression);
                             break;
                         }
-                    case "datein":
-                    case "datenotin":
+                    case Op.StartsWithAny:
+                    case Op.EndsWithAny:
+                    case Op.ContainsAny:
+                        {
+                            Type collectionElement = GetElementType(constant.Type);
+                            if (collectionElement != typeof(string))
+                                throw new FrameworkException($"Generic filter operation '{filter.Operation}' is not supported for collection with element type '{collectionElement}'. Expected element type is 'string'.");
+                            var filterValues = constant.Value as IEnumerable<string>
+                                ?? throw new FrameworkException($"Generic filter operation '{filter.Operation}' is not supported for collection with type '{constant.Value?.GetType()}'. Expected a collection of 'string' elements.");
+
+                            MethodInfo dbMethod = filterOperation switch
+                            {
+                                Op.StartsWithAny => _ormUtility.StartsWithCaseInsensitiveMethod,
+                                Op.EndsWithAny => _ormUtility.EndsWithCaseInsensitiveMethod,
+                                _ => _ormUtility.ContainsCaseInsensitiveMethod
+                            };
+                            Expression stringMemberAccess = GetStringMemberAccess(memberAccess, propertyBasicType, filter.Operation);
+                            var filterExpressions = filterValues
+                                .Select(filterValue => Expression.Call(dbMethod, stringMemberAccess, Expression.Constant(filterValue, typeof(string))))
+                                .ToList();
+
+                            if (filterExpressions.Count == 0)
+                                return Expression.Lambda(Expression.Constant(false), parameter); // If no patterns are provided, no items should be returned.
+                            expression = CombineToOrExpressions(filterExpressions);
+                            break;
+                        }
+                    case Op.DateIn:
+                    case Op.DateNotIn:
                         {
                             if (propertyBasicType != typeof(DateTime))
                                 throw new FrameworkException("Generic filter operation '" + filter.Operation
@@ -349,12 +381,12 @@ namespace Rhetos.Dom.DefaultConcepts
                                 Expression.GreaterThanOrEqual(memberAccess, Expression.Constant(date1, typeof(DateTime?))),
                                 Expression.LessThan(memberAccess, Expression.Constant(date2, typeof(DateTime?))));
 
-                            if (filter.Operation.Equals("datenotin", StringComparison.OrdinalIgnoreCase))
+                            if (filterOperation == Op.DateNotIn)
                                 expression = Expression.Not(expression);
                             break;
                         }
-                    case "in":
-                    case "notin":
+                    case Op.In:
+                    case Op.NotIn:
                         {
                             Type collectionElement = GetElementType(constant.Type);
                             Expression convertedMemberAccess = memberAccess.Type != collectionElement
@@ -369,7 +401,7 @@ namespace Rhetos.Dom.DefaultConcepts
 
                             expression = _ormUtility.OptimizeContains(Expression.Call(containsMethod, _ormUtility.CreateContainsItemsExpression(constant.Value), convertedMemberAccess));
 
-                            if (filter.Operation.Equals("notin", StringComparison.OrdinalIgnoreCase))
+                            if (filterOperation == Op.NotIn)
                                 expression = Expression.Not(expression);
                             break;
                         }
@@ -381,6 +413,28 @@ namespace Rhetos.Dom.DefaultConcepts
             }
 
             return Expression.Lambda(resultCondition, parameter);
+        }
+
+        private static Expression CombineToOrExpressions(IEnumerable<MethodCallExpression> filterExpressions)
+        {
+            Expression expression = null;
+            foreach (var fe in filterExpressions)
+                expression = expression == null ? fe : Expression.OrElse(expression, fe);
+            if (expression == null)
+                throw new FrameworkException("Unexpected filter expressions without elements.");
+            return expression;
+        }
+
+        /// <summary>Returns <paramref name="memberAccess"/> if it is a string property, otherwise casts it to string.
+        /// This enables support for generic filter operations such as 'StartsWith' on other property types besides string, for example on integer property.</summary>
+        private Expression GetStringMemberAccess(Expression memberAccess, Type propertyBasicType, string debugFilterOperation)
+        {
+            if (propertyBasicType == typeof(string))
+                return memberAccess;
+            else if (propertyBasicType == typeof(int))
+                return Expression.Call(_ormUtility.CastToStringMethod, memberAccess);
+            else    
+                throw new FrameworkException("Generic filter operation '" + debugFilterOperation + "' is not supported on property type '" + propertyBasicType.Name + "'.");
         }
 
         private static Type GetElementType(IEnumerable enumerable)